requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
bcrypt==4.3.0
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.5.2
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cookiecutter==2.4.0
cryptography==44.0.2
cycler==0.12.1
distlib==0.3.7
filelock==3.13.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flet==0.12.1
flet-core==0.12.1
flet-runtime==0.12.1
fonttools==4.45.1
google-auth==2.40.0
gunicorn==22.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mysql-connector-python==8.2.0
numpy==1.26.1
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.21.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.16.1
PyJWT==2.10.1
PyMySQL==1.0.2
pyparsing==3.1.1
pypng==0.20220715.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.6.0
rsa==4.9.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.24.6
watchdog==3.0.0
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
